refactor(server): apply the agent session review feedback to nango_proxy
Three things earlier meta tool PRs were pulled up on, applied here so we
don't repeat them.

Meta tools are now declared once, keyed by the session field, and the
accepted keys, the defaults and the created reply are derived from that
table. Wiring a meta tool into some of them and missing others is a
compile error now.

Moves the shared proxy execution into mcpProxy.service.ts. The agent
session tool was importing it out of the management MCP controller,
where a tool importing another tool's folder is the layering we already
agreed to avoid. Both tools now import the service, like tool search
does.

Also drops a test that asserted only that validation said something, and
covers the parse and summary helpers.
