+++ packages/server/lib/controllers/agent/mcp/proxy/proxy.ts
@@ -2,14 +2,12 @@ import tracer from 'dd-trace';
 
 import { Err } from '@nangohq/utils';
 
-import { executeMcpProxyRequest } from '../../../mcp/proxy/execute.js';
-import { MAX_MCP_PROXY_RESPONSE_SIZE_LABEL } from '../../../mcp/proxy/response.js';
-import { proxyRequestOutputSchema } from '../../../mcp/proxy/schema.js';
+import { executeMcpProxyRequest, MAX_MCP_PROXY_RESPONSE_SIZE_LABEL, mcpProxyResponseSchema } from '../../../../services/mcpProxy.service.js';
 import { PublicMcpError } from '../../../mcp/utils.js';
 import { defineAgentSessionMcpTool } from '../sessionTool.js';
 import { proxyInputSchema } from './schema.js';
 
-import type { ProxyRequestOutput } from '../../../mcp/proxy/schema.js';
+import type { McpProxyResponse } from '../../../../services/mcpProxy.service.js';
 import type { Result } from '@nangohq/utils';
 import type { Span } from 'dd-trace';
 
@@ -17,7 +15,7 @@ export const proxyTool = defineAgentSessionMcpTool({
     name: 'nango_proxy',
     description: `Make an authenticated HTTP request to a provider API, on the connection this session resolved for the integration. The escape hatch for when no tool covers what you need, so search with nango_tool_search first and use this only if nothing fits. Returns JSON or UTF-8 text responses up to ${MAX_MCP_PROXY_RESPONSE_SIZE_LABEL}.`,
     inputSchema: proxyInputSchema,
-    outputSchema: proxyRequestOutputSchema,
+    outputSchema: mcpProxyResponseSchema,
     annotations: {
         readOnlyHint: false,
         destructiveHint: true,
@@ -39,7 +37,7 @@ export const proxyTool = defineAgentSessionMcpTool({
             return Err(new PublicMcpError(`Integration '${integrationId}' has no connection in this session.`));
         }
 
-        return await tracer.trace<Promise<Result<ProxyRequestOutput>>>('server.mcp.agentSession.proxy', async (span: Span) => {
+        return await tracer.trace<Promise<Result<McpProxyResponse>>>('server.mcp.agentSession.proxy', async (span: Span) => {
             span.setTag('nango.agentSessionId', session.id)
                 .setTag('nango.accountId', account.id)
                 .setTag('nango.environmentId', environment.id)
