+++ packages/server/lib/services/agentSessionCreation.service.ts
@@ -38,10 +38,19 @@ const EXPIRES_IN_UNITS_IN_MS: Record<string, number> = {
     d: 24 * 60 * 60 * 1000
 };
 
-const META_TOOLS = ['nango_tool_search', 'nango_execute', 'nango_proxy'] as const;
+/**
+ * The one place a meta tool is declared. The accepted keys, the defaults and the created reply are
+ * all derived from this, so a new meta tool cannot be wired into some of them and missed in others.
+ * Keying it by the session field makes leaving a field unset a compile error.
+ */
+const META_TOOLS = {
+    nangoToolSearch: { name: 'nango_tool_search', enabledByDefault: true },
+    nangoExecute: { name: 'nango_execute', enabledByDefault: true },
+    // Off by default: it reaches any endpoint of a connected integration, not only the toolset's tools.
+    nangoProxy: { name: 'nango_proxy', enabledByDefault: false }
+} as const satisfies Record<keyof AgentSessionMetaTools, { name: keyof AgentSessionMetaToolsSummary; enabledByDefault: boolean }>;
 
-/** nango_proxy is off by default: it reaches any endpoint of a connected integration, not only the toolset's tools. */
-const DEFAULT_META_TOOLS: AgentSessionMetaTools = { nangoToolSearch: true, nangoExecute: true, nangoProxy: false };
+const META_TOOL_NAMES = Object.values(META_TOOLS).map((metaTool) => metaTool.name);
 
 export const agentSessionExpiresInSchema = z
     .string()
@@ -175,7 +184,7 @@ async function runCreation(params: CreateAgentSessionParams, logCtx: LogContextO
         return Err(
             new AgentSessionCreationError({
                 code: 'unknown_meta_tool',
-                message: `${metaTools.unknown.length} ${metaTools.unknown.length === 1 ? 'key is' : 'keys are'} not a meta tool Nango ships. Supported meta tools are ${META_TOOLS.join(', ')}.`,
+                message: `${metaTools.unknown.length} ${metaTools.unknown.length === 1 ? 'key is' : 'keys are'} not a meta tool Nango ships. Supported meta tools are ${META_TOOL_NAMES.join(', ')}.`,
                 payload: { meta_tools: metaTools.unknown }
             })
         );
@@ -234,25 +243,28 @@ async function runCreation(params: CreateAgentSessionParams, logCtx: LogContextO
         token: token.value.token,
         mcpUrl: `${baseUrl}/session/${session.value.id}/mcp`,
         toolset: toolsetSummary(session.value.compiledToolset, session.value.resolvedConnections),
-        metaTools: {
-            nango_tool_search: session.value.metaTools.nangoToolSearch,
-            nango_execute: session.value.metaTools.nangoExecute,
-            nango_proxy: session.value.metaTools.nangoProxy
-        }
+        metaTools: metaToolsSummary(session.value.metaTools)
     });
 }
 
-function parseMetaTools(requested: Record<string, boolean> | undefined): { applied: AgentSessionMetaTools; unknown: string[] } {
-    const unknown = Object.keys(requested ?? {}).filter((key) => !META_TOOLS.includes(key as (typeof META_TOOLS)[number]));
+export function parseMetaTools(requested: Record<string, boolean> | undefined): { applied: AgentSessionMetaTools; unknown: string[] } {
+    const unknown = Object.keys(requested ?? {}).filter((key) => !META_TOOL_NAMES.includes(key as keyof AgentSessionMetaToolsSummary));
 
-    return {
-        applied: {
-            nangoToolSearch: requested?.['nango_tool_search'] ?? DEFAULT_META_TOOLS.nangoToolSearch,
-            nangoExecute: requested?.['nango_execute'] ?? DEFAULT_META_TOOLS.nangoExecute,
-            nangoProxy: requested?.['nango_proxy'] ?? DEFAULT_META_TOOLS.nangoProxy
-        },
-        unknown
-    };
+    const applied = {} as Record<keyof AgentSessionMetaTools, boolean>;
+    for (const [field, metaTool] of Object.entries(META_TOOLS) as [keyof AgentSessionMetaTools, (typeof META_TOOLS)[keyof AgentSessionMetaTools]][]) {
+        applied[field] = requested?.[metaTool.name] ?? metaTool.enabledByDefault;
+    }
+
+    return { applied, unknown };
+}
+
+export function metaToolsSummary(metaTools: AgentSessionMetaTools): AgentSessionMetaToolsSummary {
+    const summary = {} as AgentSessionMetaToolsSummary;
+    for (const [field, metaTool] of Object.entries(META_TOOLS) as [keyof AgentSessionMetaTools, (typeof META_TOOLS)[keyof AgentSessionMetaTools]][]) {
+        summary[metaTool.name] = metaTools[field];
+    }
+
+    return summary;
 }
 
 function rejected(error: { code: AgentSessionCreationErrorCode; message: string; payload: Record<string, unknown> }): AgentSessionCreationError {
