+++ packages/server/lib/services/agentSessionCreation.service.unit.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { expiresInToMs, toolsetSummary } from './agentSessionCreation.service.js';
+import { expiresInToMs, metaToolsSummary, parseMetaTools, toolsetSummary } from './agentSessionCreation.service.js';
 
 describe('expiresInToMs', () => {
     it.each([
@@ -36,6 +36,39 @@ describe('toolsetSummary', () => {
         expect(summary).toStrictEqual({
             notion: { connected: true, tools_pinned: 1, tools_searchable: 1 },
             reddit: { connected: false, tools_pinned: 0, tools_searchable: 1 }
+        });
+    });
+});
+
+describe('parseMetaTools', () => {
+    it('applies the defaults when nothing is requested', () => {
+        expect(parseMetaTools(undefined)).toStrictEqual({
+            applied: { nangoToolSearch: true, nangoExecute: true, nangoProxy: false },
+            unknown: []
+        });
+    });
+
+    it('overrides only the meta tools the caller named', () => {
+        expect(parseMetaTools({ nango_execute: false, nango_proxy: true })).toStrictEqual({
+            applied: { nangoToolSearch: true, nangoExecute: false, nangoProxy: true },
+            unknown: []
+        });
+    });
+
+    it('collects the keys that are not meta tools Nango ships', () => {
+        const parsed = parseMetaTools({ nango_proxy: true, nango_teleport: true, proxy: false });
+
+        expect(parsed.unknown).toStrictEqual(['nango_teleport', 'proxy']);
+        expect(parsed.applied.nangoProxy).toBe(true);
+    });
+});
+
+describe('metaToolsSummary', () => {
+    it('reports every meta tool under its request name', () => {
+        expect(metaToolsSummary({ nangoToolSearch: true, nangoExecute: false, nangoProxy: true })).toStrictEqual({
+            nango_tool_search: true,
+            nango_execute: false,
+            nango_proxy: true
         });
     });
 });
